# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Farmers Marketplace API
  description: REST API for the Mysore Farmer Marketplace — powered by Google Cloud Functions and Cloud Firestore.
  version: 1.0.0
servers:
  - url: https://asia-south1-mysore-farmer-marketplace.cloudfunctions.net/farmers-marketplace-api
    description: Production (Cloud Functions)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/about:
    get:
      summary: Get About Us content
      operationId: getAboutUs
      responses:
        '200':
          description: About Us content
    put:
      summary: Update About Us content (admin only)
      operationId: updateAboutUs
      security:
        - mockAuth: []
      responses:
        '200':
          description: Content updated
...
components:
  securitySchemes:
    mockAuth:
      type: apiKey
      in: header
      name: X-User-Id
      description: Mock authentication via custom headers (X-User-Id, X-User-Role, etc.)